+++ NEWS.md
@@ -1,4 +1,4 @@
-# attachment 0.2.2
+# attachment 0.2.3
 
 ## Major changes